marketing_content_generator/marketing_content_generator/requirements.txt
openai
chromadb
python-dotenv
tqdm